@@ -80,55 +81,58 @@
     "lottie-web": "^5.8.1",
     "react": "^17.0.0",
     "react-dom": "^17.0.0",
+    "reconnecting-websocket": "^4.4.0",
     "reduct": "^3.3.1",
     "supertokens-website": "^11.0.2",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@babel/core": "7.18.5",
-    "@babel/plugin-proposal-class-properties": "7.17.12",
-    "@babel/plugin-proposal-decorators": "7.18.2",
-    "@babel/plugin-proposal-object-rest-spread": "7.18.0",
-    "@babel/preset-env": "7.18.2",
-    "@babel/preset-typescript": "7.17.12",
-    "@commitlint/cli": "17.0.2",
-    "@commitlint/config-conventional": "17.0.2",
-    "@swc/core": "1.2.203",
+    "@abraham/reflection": "0.10.0",
+    "@babel/core": "7.18.6",
+    "@babel/plugin-proposal-class-properties": "7.18.6",
+    "@babel/plugin-proposal-decorators": "7.18.6",
+    "@babel/plugin-proposal-object-rest-spread": "7.18.6",
+    "@babel/preset-env": "7.18.6",
+    "@babel/preset-react": "7.18.6",
+    "@babel/preset-typescript": "7.18.6",
+    "@commitlint/cli": "17.0.3",
+    "@commitlint/config-conventional": "17.0.3",
+    "@swc/core": "1.2.208",
     "@swc/jest": "0.2.21",
     "@types/debug": "4.1.7",
     "@types/filesystem": "0.0.32",
     "@types/firefox": "0.0.31",
     "@types/get-port": "4.2.0",
     "@types/har-format": "1.2.8",
-    "@types/jest": "28.1.1",
+    "@types/jest": "28.1.3",
     "@types/jsonwebtoken": "8.5.8",
     "@types/long": "4.0.2",
-    "@types/node": "16.11.41",
+    "@types/node": "16.11.43",
     "@types/node-fetch": "2.6.2",
     "@types/react": "17.0.47",
     "@types/react-dom": "17.0.17",
     "@types/uuid": "8.3.4",
-    "@typescript-eslint/eslint-plugin": "5.28.0",
-    "@typescript-eslint/parser": "5.28.0",
+    "@typescript-eslint/eslint-plugin": "5.30.4",
+    "@typescript-eslint/parser": "5.30.4",
     "@webexts/build-utils": "^0.0.0",
     "@wext/shipit": "0.2.1",
     "@yarnpkg/pnpify": "3.1.3",
     "JSON2016": "1.0.0",
-    "addons-linter": "5.8.0",
+    "addons-linter": "5.10.0",
     "babel-plugin-parameter-decorator": "1.0.16",
     "babel-plugin-transform-typescript-metadata": "0.3.2",
     "copy-webpack-plugin": "11.0.0",
     "crx-hotreload": "1.0.6",
-    "eslint": "8.17.0",
+    "eslint": "8.19.0",
     "eslint-config-prettier": "8.5.0",
     "eslint-import-resolver-node": "0.3.6",
     "eslint-plugin-import": "2.26.0",
     "eslint-plugin-jest": "26.5.3",
-    "eslint-plugin-jsx-a11y": "6.5.1",
+    "eslint-plugin-jsx-a11y": "6.6.0",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-prettier": "4.0.0",
+    "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-promise": "6.0.0",
-    "eslint-plugin-react": "7.30.0",
+    "eslint-plugin-react": "7.30.1",
     "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-standard": "5.0.0",
     "foxr": "0.10.1",
