@@ -51,60 +51,62 @@
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@babel/core": "7.18.5",
-    "@babel/plugin-proposal-class-properties": "7.17.12",
-    "@babel/plugin-proposal-decorators": "7.18.2",
-    "@babel/plugin-proposal-object-rest-spread": "7.18.0",
-    "@babel/preset-env": "7.18.2",
-    "@babel/preset-typescript": "7.17.12",
+    "@abraham/reflection": "0.10.0",
+    "@babel/core": "7.18.6",
+    "@babel/plugin-proposal-class-properties": "7.18.6",
+    "@babel/plugin-proposal-decorators": "7.18.6",
+    "@babel/plugin-proposal-object-rest-spread": "7.18.6",
+    "@babel/preset-env": "7.18.6",
+    "@babel/preset-react": "7.18.6",
+    "@babel/preset-typescript": "7.18.6",
     "@coil/puppeteer-utils": "^0.0.0",
     "@coil/webpack-utils": "0.0.0",
-    "@commitlint/cli": "17.0.2",
-    "@commitlint/config-conventional": "17.0.2",
-    "@swc/core": "1.2.203",
+    "@commitlint/cli": "17.0.3",
+    "@commitlint/config-conventional": "17.0.3",
+    "@swc/core": "1.2.208",
     "@swc/jest": "0.2.21",
-    "@types/jest": "28.1.1",
+    "@types/jest": "28.1.3",
     "@types/long": "4.0.2",
-    "@types/node": "16.11.41",
+    "@types/node": "16.11.43",
     "@types/node-fetch": "2.6.2",
     "@types/uuid": "8.3.4",
-    "@typescript-eslint/eslint-plugin": "5.28.0",
-    "@typescript-eslint/parser": "5.28.0",
+    "@typescript-eslint/eslint-plugin": "5.30.4",
+    "@typescript-eslint/parser": "5.30.4",
     "@yarnpkg/pnpify": "3.1.3",
     "JSON2016": "1.0.0",
     "babel-plugin-parameter-decorator": "1.0.16",
     "babel-plugin-transform-typescript-metadata": "0.3.2",
     "copy-webpack-plugin": "11.0.0",
-    "eslint": "8.17.0",
+    "eslint": "8.19.0",
     "eslint-config-prettier": "8.5.0",
     "eslint-import-resolver-node": "0.3.6",
     "eslint-plugin-import": "2.26.0",
     "eslint-plugin-jest": "26.5.3",
-    "eslint-plugin-jsx-a11y": "6.5.1",
+    "eslint-plugin-jsx-a11y": "6.6.0",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-prettier": "4.0.0",
+    "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-promise": "6.0.0",
-    "eslint-plugin-react": "7.30.0",
+    "eslint-plugin-react": "7.30.1",
     "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-standard": "5.0.0",
     "husky": "8.0.1",
     "jest": "28.1.1",
     "jest-environment-jsdom": "28.1.1",
-    "lint-staged": "13.0.2",
-    "nodemon": "2.0.16",
+    "lint-staged": "13.0.3",
+    "nodemon": "2.0.18",
     "prettier": "2.7.1",
-    "puppeteer": "14.4.0",
+    "puppeteer": "15.3.0",
     "regenerator-runtime": "0.13.9",
     "rimraf": "3.0.2",
     "ts-jest": "28.0.5",
-    "ts-loader": "9.3.0",
-    "ts-node": "10.8.1",
+    "ts-loader": "9.3.1",
+    "ts-node": "10.8.2",
     "tsconfig-paths": "4.0.0",
-    "typescript": "4.7.3",
+    "typescript": "4.7.4",
     "webpack": "5.73.0",
     "webpack-bundle-analyzer": "4.5.0",
     "webpack-cli": "4.10.0",
-    "webpack-dev-server": "4.9.2",
+    "webpack-dev-server": "4.9.3",
     "webpack-merge": "5.8.0"
   }
 }
