@@ -1,6 +1,8 @@
 import { decorate, inject, injectable } from 'inversify'
+import * as idbKv from 'idb-keyval'
 import * as anonymousTokens from '@coil/anonymous-tokens'
 import { BlindToken, CurvePoints } from '@coil/privacypass-sjcl'
+import { StorableBlindToken } from '@coil/anonymous-tokens'
 
 import * as tokens from '../../types/tokens'
 import { BuildConfig } from '../../types/BuildConfig'
@@ -11,39 +13,33 @@ decorate(injectable(), anonymousTokens.AnonymousTokens)
 
 const ANON_TOKEN_BATCH_SIZE = 10
 
-class TokenStore {
-  private storage: Storage
-  constructor(localStorage: Storage) {
-    this.storage = localStorage
-  }
+@injectable()
+export class IDBTokenStore implements anonymousTokens.TokenStore {
+  private store = idbKv.createStore('anonTokens', 'anonTokens')
 
-  async setItem(key: string, value: string): Promise<string> {
-    this.storage.setItem(key, value)
-    return Promise.resolve(value)
-  }
-
-  async removeItem(key: string): Promise<void> {
-    this.storage.removeItem(key)
-    return Promise.resolve()
+  async clear() {
+    return idbKv.clear(this.store)
   }
 
   async iterate(
-    fn: (
-      value: string,
-      key: string,
-      iterationNumber: number
-    ) => anonymousTokens.StorableBlindToken | undefined
-  ): Promise<anonymousTokens.StorableBlindToken | undefined> {
-    let i = 0
-    let key
-    while ((key = this.storage.key(i)) !== null) {
-      const value = this.storage.getItem(key)
-      if (value === null) continue
-      const result = fn(value, key, i)
-      if (result) return Promise.resolve(result)
-      i++
+    fn: (key: string, value: string) => StorableBlindToken | undefined
+  ): Promise<StorableBlindToken | undefined> {
+    const entries = await idbKv.entries<string, string>(this.store)
+    for (const [key, val] of entries) {
+      const token = fn(key, val)
+      if (token) {
+        return token
+      }
     }
-    return Promise.resolve(undefined)
+  }
+
+  async removeItem(key: string): Promise<void> {
+    return idbKv.del(key, this.store)
+  }
+
+  async setItem(key: string, value: string): Promise<string> {
+    await idbKv.set(key, value, this.store)
+    return value
   }
 }
 
@@ -52,13 +48,13 @@ export class AnonymousTokens extends anonymousTokens.AnonymousTokens {
   constructor(
     @inject(tokens.BuildConfig) private buildConfig: BuildConfig,
     @inject(tokens.CoilDomain) coilHost: string,
-    @inject(Storage) storage: Storage,
+    @inject(tokens.TokenStore) store: anonymousTokens.TokenStore,
     @logger('AnonymousTokens') debug: Logger
   ) {
     super({
       redeemerUrl: coilHost + '/redeemer',
       signerUrl: coilHost + '/issuer',
-      store: new TokenStore(storage),
+      store,
       debug,
       batchSize: ANON_TOKEN_BATCH_SIZE
     })
