@@ -1,10 +1,8 @@
 import { inject, injectable } from 'inversify'
 import { GraphQlClient } from '@coil/client'
-import { MonetizationState } from '@webmonetization/types'
-import { resolvePaymentEndpoint } from '@webmonetization/polyfill-utils'
 
 import { notNullOrUndef } from '../../util/nullables'
-import { StorageService } from '../../services/storage'
+import { StoreService } from '../../services/storage'
 import * as tokens from '../../types/tokens'
 import {
   AdaptedSite,
@@ -12,27 +10,18 @@ import {
   CheckIFrameIsAllowedFromBackground,
   ClosePopup,
   ContentScriptInit,
-  MonetizationProgress,
-  MonetizationStart,
+  ContentScriptInitResponse,
   OnFrameAllowedChanged,
-  PauseWebMonetization,
   ReportCorrelationIdFromIFrameContentScript,
   ReportCorrelationIdToParentContentScript,
-  ResumeWebMonetization,
   SetMonetizationState,
-  SetStreamControls,
-  StartWebMonetization,
   ToBackgroundMessage
 } from '../../types/commands'
-import { LocalStorageProxy } from '../../types/storage'
-import { TabState } from '../../types/TabState'
-import { getFrameSpec, getTab } from '../../util/tabs'
+import { getFrameSpec } from '../../util/tabs'
 import { FrameSpec } from '../../types/FrameSpec'
 import { BuildConfig } from '../../types/BuildConfig'
-import { User } from '../../types/user'
 
 import { MultipleInstanceDetector } from './multipleInstanceDetector'
-import { StreamMoneyEvent } from './Stream'
 import { AuthService } from './AuthService'
 import { TabStates } from './TabStates'
 import { Streams } from './Streams'
@@ -41,10 +30,14 @@ import { Logger, logger } from './utils'
 import { YoutubeService } from './YoutubeService'
 import { BackgroundFramesService } from './BackgroundFramesService'
 import { TippingService } from './TippingService'
-import { StreamAssociations } from './StreamAssociations'
 import { ActiveTabLogger } from './ActiveTabLogger'
+import { StreamAssociations } from './StreamAssociations'
+import { SPSPState } from './SPSPState'
+import { MonetizationService } from './MonetizationService'
 
 import MessageSender = chrome.runtime.MessageSender
+
+import { WM2OriginTrial } from './WM2OriginTrial'
 
 @injectable()
 export class BackgroundScript {
@@ -52,13 +45,14 @@ export class BackgroundScript {
     private popup: PopupBrowserAction,
     private assoc: StreamAssociations,
     private streams: Streams,
+    private spspState: SPSPState,
     private tabStates: TabStates,
-    private storage: StorageService,
-    @inject(tokens.LocalStorageProxy)
-    private store: LocalStorageProxy,
+    private storage: StoreService,
     private auth: AuthService,
+    private monetization: MonetizationService,
     private tippingService: TippingService,
     private youtube: YoutubeService,
+    private wm2OriginTrial: WM2OriginTrial,
     private activeTabLogger: ActiveTabLogger,
     private framesService: BackgroundFramesService,
     @inject(tokens.LoggingEnabled)
@@ -102,8 +96,9 @@ export class BackgroundScript {
     this.setTabsOnRemovedListener()
     this.setFramesOnChangedListener()
     this.setFramesOnRemovedListener()
-    this.routeStreamsMoneyEventsToContentScript()
-    this.handleStreamsAbortEvent()
+
+    this.monetization.init()
+
     this.popup.setDefaultInactive()
     this.framesService.monitor()
     this.bindOnInstalled()
@@ -125,7 +120,7 @@ export class BackgroundScript {
         this.log('tabs.onActivated %o', activeInfo)
       }
       this.activeTab = activeInfo.tabId
-      this.reloadTabState({ from: 'onActivated' })
+      this.tabStates.reloadTabState({ from: 'onActivated' })
     })
     if (this.buildConfig.logTabsApiEvents) {
       if (this.api.tabs.onActiveChanged) {
@@ -155,13 +150,7 @@ export class BackgroundScript {
         const close: ClosePopup = {
           command: 'closePopup'
         }
-        // Storage events are only fired if the value actually changes, not
-        // on every localStorage `setItem` call. So we set the first 16
-        // characters of the stored value to the current time.
-        this.storage.setRaw(
-          '$$popupCommand',
-          Date.now().toString().padStart(16, '0') + JSON.stringify(close)
-        )
+        void this.api.runtime.sendMessage(close)
         this.api.tabs.query({ active: true, currentWindow: true }, tabs => {
           if (this.api.runtime.lastError) {
             // In this case we clicked on a non-active tab as the focusing
@@ -174,7 +163,7 @@ export class BackgroundScript {
             return
           }
           this.activeTab = tabs[0].id
-          this.reloadTabState({ from: 'onFocusChanged' })
+          this.tabStates.reloadTabState({ from: 'onFocusChanged' })
         })
       })
     }
@@ -198,11 +187,7 @@ export class BackgroundScript {
 
   private initializeActiveTab() {
     // initialize tab so we can set its state right away
-    this.api.tabs.query({ active: true, currentWindow: true }, tabs => {
-      if (tabs.length === 0 || tabs[0].id == null) return
-      this.activeTab = tabs[0].id
-      this.reloadTabState({ from: 'initial' })
-    })
+    this.tabStates.reloadTabState({ from: 'initial' })
   }
 
   private setTabsOnRemovedListener() {
@@ -287,95 +272,8 @@ export class BackgroundScript {
     })
   }
 
-  private routeStreamsMoneyEventsToContentScript() {
-    // pass stream monetization events to the correct tab
-    this.streams.on('money', (details: StreamMoneyEvent) => {
-      const frame = this.assoc.getFrame(details.requestId)
-      const { tabId, frameId } = frame
-      if (details.packetNumber === 0) {
-        const message: MonetizationStart = {
-          command: 'monetizationStart',
-          data: {
-            paymentPointer: details.paymentPointer,
-            requestId: details.requestId
-          }
-        }
-        this.api.tabs.sendMessage(tabId, message, { frameId })
-      }
-
-      const message: MonetizationProgress = {
-        command: 'monetizationProgress',
-        data: {
-          paymentPointer: details.paymentPointer,
-          amount: details.amount,
-          assetCode: details.assetCode,
-          requestId: details.requestId,
-          assetScale: details.assetScale,
-          sentAmount: details.sentAmount,
-          receipt: details.receipt
-        }
-      }
-      this.handleMonetizedSite(frame, details.initiatingUrl, details)
-      this.api.tabs.sendMessage(tabId, message, { frameId })
-      this.savePacketToHistoryDb(details)
-    })
-  }
-
-  private savePacketToHistoryDb(_: StreamMoneyEvent) {
-    // this.db.incrementSite(details)
-  }
-
   async adaptedPageDetails(variables: { url: string; channelId?: string }) {
     return this.client.adaptedPage(variables.url, variables.channelId)
-  }
-
-  private async setBrowserActionStateFromAuthAndTabState() {
-    const isAuthenticated = await this.auth.isAuthenticated()
-
-    if (!isAuthenticated) {
-      this.popup.disable()
-    }
-
-    const tabId = this.activeTab
-
-    if (tabId) {
-      const tabState = this.tabStates.getActiveOrDefault()
-
-      if (Object.values(tabState.frameStates).find(f => f.monetized)) {
-        this.tabStates.setIcon(tabId, 'monetized')
-      }
-
-      if (!isAuthenticated) {
-        this.tabStates.setIcon(tabId, 'unavailable')
-      } else {
-        this.popup.enable()
-
-        const tabState = this.tabStates.getActiveOrDefault()
-        const frameStates = Object.values(tabState.frameStates)
-        const hasStream = frameStates.find(f => f.monetized)
-        const hasBeenPaid = hasStream && frameStates.find(f => f.total > 0)
-
-        if (hasStream) {
-          this.tabStates.setIcon(tabId, 'monetized')
-          if (hasBeenPaid) {
-            const state =
-              tabState.playState === 'playing'
-                ? 'streaming'
-                : 'streaming-paused'
-            this.tabStates.setIcon(tabId, state)
-          } else {
-            // Need to check if the user is able to tip with the new ui.
-            // This assumes that the site is monetized and
-            // that the user meets certain criteria
-            if (this.tippingService.userCanTip()) {
-              this.tabStates.setIcon(tabId, 'tipping')
-            }
-          }
-        } else {
-          this.tabStates.setIcon(tabId, 'inactive')
-        }
-      }
-    }
   }
 
   async handleMessage(
@@ -385,6 +283,15 @@ export class BackgroundScript {
     sendResponse: (response: any) => any
   ) {
     switch (request.command) {
+      case 'storeGetItems':
+        sendResponse({ items: this.storage.items() })
+        break
+      case 'storeSetItem':
+        this.storage.set(request.data.key, request.data.value)
+        break
+      case 'storeRemoveItem':
+        this.storage.remove(request.data.key)
+        break
       case 'log':
         this.log('log command:', request.data)
         sendResponse(true)
@@ -407,25 +314,24 @@ export class BackgroundScript {
         break
       case 'startWebMonetization':
         this.log('got startwebmonetization')
-        sendResponse(await this.startWebMonetization(request, sender))
+        sendResponse(
+          await this.monetization.startWebMonetization(request, sender)
+        )
         break
       case 'pauseWebMonetization':
-        sendResponse(this.pauseWebMonetization(request, sender))
+        sendResponse(this.monetization.pauseWebMonetization(request, sender))
         break
       case 'resumeWebMonetization':
-        sendResponse(this.resumeWebMonetization(request, sender))
+        sendResponse(this.monetization.resumeWebMonetization(request, sender))
         break
       case 'stopWebMonetization':
-        sendResponse(this.stopWebMonetization(sender))
+        sendResponse(this.monetization.stopWebMonetization(request, sender))
         break
       case 'isRateLimited':
         sendResponse(await this.isRateLimited())
         break
-      case 'setStreamControls':
-        sendResponse(this.setStreamControls(request, sender))
-        break
       case 'contentScriptInit':
-        sendResponse(this.contentScriptInit(request, sender))
+        sendResponse(await this.contentScriptInit(request, sender))
         break
       case 'fetchYoutubeChannelId':
         sendResponse(await this.youtube.fetchChannelId(request.data.youtubeUrl))
@@ -473,7 +379,7 @@ export class BackgroundScript {
       coilSite: url
     })
     if (url) {
-      this.reloadTabState({ from: 'setCoilUrlForPopupIfNeeded' })
+      this.tabStates.reloadTabState({ from: 'setCoilUrlForPopupIfNeeded' })
     }
   }
 
@@ -495,99 +401,11 @@ export class BackgroundScript {
     return newest
   }
 
-  setFrameMonetized(
-    { tabId, frameId }: FrameSpec,
-    senderUrl: string,
-    total?: number
-  ) {
-    const tabState = this.tabStates.get(tabId)
-
-    this.tabStates.setFrame(
-      { tabId, frameId },
-      {
-        monetized: true,
-        total: total || (tabState.frameStates[frameId]?.total ?? 0)
-      }
-    )
-
-    if (this.activeTab === tabId) {
-      this.reloadTabState()
-    }
-  }
-
-  mayMonetizeSite(sender: chrome.runtime.MessageSender, initiatingUrl: string) {
-    this.setFrameMonetized(getFrameSpec(sender), initiatingUrl)
-  }
-
-  handleMonetizedSite(
-    { tabId, frameId }: FrameSpec,
-    url: string,
-    packet: { sentAmount: string }
-  ) {
-    const tabState = this.tabStates.get(tabId)
-    const frameTotal = tabState?.frameStates[frameId]?.total ?? 0
-    const newFrameTotal = frameTotal + Number(packet?.sentAmount ?? 0)
-    this.setFrameMonetized({ tabId, frameId }, url, newFrameTotal)
-  }
-
   adaptedSite(data: AdaptedSite['data'], sender: MessageSender) {
     const { spec } = getFrameSpec(sender)
     this.tabStates.setFrame(spec, {
       adapted: data.state
     })
-  }
-
-  async reloadTabState(opts: { from?: string } = {}) {
-    const { from } = opts
-
-    const tab = this.activeTab
-    const state = () => this.tabStates.get(tab)
-    this.setLocalStorageFromState(state())
-    await this.setBrowserActionStateFromAuthAndTabState()
-    // Don't work off stale state, set(...) creates a copy ...
-    this.popup.setBrowserAction(tab, state())
-    if (from) {
-      this.log(
-        `reloadTabState tab=${tab}`,
-        `from=${from}`,
-        JSON.stringify(state(), null, 2)
-      )
-    }
-  }
-
-  private setLocalStorageFromState(state: TabState) {
-    const frameStates = Object.values(state.frameStates)
-
-    state && state.coilSite
-      ? this.storage.set('coilSite', state.coilSite)
-      : this.storage.remove('coilSite')
-    // TODO: Another valid case might be a singular adapted iframe inside a non
-    // monetized top page.
-    this.storage.set('adapted', Boolean(state?.frameStates[0]?.adapted))
-    state && frameStates.find(f => f.monetized)
-      ? this.storage.set('monetized', true)
-      : this.storage.remove('monetized')
-
-    if (state && state.playState && state.stickyState) {
-      this.store.playState = state.playState
-      this.store.stickyState = state.stickyState
-    } else if (state) {
-      delete this.store.playState
-      delete this.store.stickyState
-    }
-
-    if (this.buildConfig.extensionBuildString) {
-      this.store.extensionBuildString = this.buildConfig.extensionBuildString
-    }
-    if (this.buildConfig.extensionPopupFooterString) {
-      this.store.extensionPopupFooterString =
-        this.buildConfig.extensionPopupFooterString
-    }
-
-    if (state) {
-      const total = frameStates.reduce((acc, val) => acc + val.total, 0)
-      this.storage.set('monetizedTotal', total)
-    }
   }
 
   async checkIFrameIsAllowedFromIFrameContentScript(sender: MessageSender) {
@@ -653,159 +471,17 @@ export class BackgroundScript {
     this.api.tabs.sendMessage(frame.tabId, message, { frameId: parentId })
   }
 
-  async startWebMonetization(
-    request: StartWebMonetization,
-    sender: MessageSender
-  ) {
-    const frame = getFrameSpec(sender)
-    const { tabId, frameId } = frame
-    const { requestId } = request.data
-
-    this.activeTabLogger.log(`startWM called with ${requestId}`, frame)
-    this.tabStates.setFrame(frame, {
-      paymentPointer: request.data.paymentPointer
-    })
-    this.tabStates.logLastMonetizationCommand(frame, 'start', requestId)
-
-    // This used to be sent from content script as a separate message
-    this.mayMonetizeSite(sender, request.data.initiatingUrl)
-
-    // This may throw so do after mayMonetizeSite has had a chance to set
-    // the page as being monetized (or attempted to be)
-    const spspEndpoint = resolvePaymentEndpoint(request.data.paymentPointer)
-
-    const userBeforeReAuth = this.store.user
-    let emittedPending = false
-    const emitPending = () => {
-      // Set the requestId so that DocumentMonetization#setState will ignore
-      // this message if tags are added and removed extremely fast.
-      this.sendSetMonetizationStateMessage(
-        frame,
-        'pending',
-        request.data.requestId
-      )
-      emittedPending = true
-    }
-
-    const setUnavailable = (fromNo: 'token' | 'subscription') => {
-      this.tabStates.setIcon(tabId, 'unavailable')
-      this.reloadTabState({ from: `no ${fromNo}` })
-    }
-
-    // If we are optimistic we have an active subscription (things could have
-    // changed since our last cached whoami query), emit pending immediately,
-    // otherwise wait until recheck auth/whoami, potentially not even emitting.
-    if (userBeforeReAuth?.subscription?.active) {
-      emitPending()
-    } else {
-      setUnavailable('subscription')
-    }
-
-    this.log('startWebMonetization, request', request)
-
-    this.log('loading token for monetization', requestId)
-    const isAuthenticated = await this.auth.isAuthenticated()
-    if (!isAuthenticated) {
-      // not signed in.
-      // eslint-disable-next-line no-console
-      if (this.loggingEnabled) {
-        console.warn('startWebMonetization cancelled; not authenticated')
-      }
-      this.activeTabLogger.log(
-        'startWebMonetization cancelled; not authenticated'
-      )
-      this.sendSetMonetizationStateMessage(
-        frame,
-        'stopped',
-        request.data.requestId
-      )
-      setUnavailable('token')
-      return false
-    }
-    if (!this.store.user?.subscription?.active) {
-      if (this.loggingEnabled) {
-        console.warn('startWebMonetization cancelled; no active subscription')
-      }
-      this.activeTabLogger.log(
-        'startWebMonetization cancelled; no active subscription',
-        frame
-      )
-      this.sendSetMonetizationStateMessage(
-        frame,
-        'stopped',
-        request.data.requestId
-      )
-      setUnavailable('subscription')
-      return false
-    }
-
-    // Check that this startWebMonetization invocation is still valid before
-    // we go ahead. Any operation that we `await`d on could have potentially
-    // masked state changes. e.g. `getTokenMaybeRefreshAndStoreState`
-    // (which will update `whoami`) which takes longer than it does to switch
-    // out a monetization tag.
-    if (
-      this.tabStates.getFrameOrDefault(frame).lastMonetization.requestId !==
-      requestId
-    ) {
-      // The pending message (if sent) will have been ignored on the content
-      // script side in this case too, so there's no need to send a stop, as
-      // will already have been stopped (for that id).
-      // If we sent a stopped message, it would need to be tagged with
-      // requestId, and would only ever be ignored (
-      // due to the new monetization request implied in the if condition)
-      this.activeTabLogger.log(
-        `startWebMonetization aborted; stale requestId: ${requestId}`,
-        frame
-      )
-      return false
-    }
-
-    if (!emittedPending) {
-      emitPending()
-    }
-
-    const lastCommand =
-      this.tabStates.getFrameOrDefault(frame).lastMonetization.command
-    if (lastCommand !== 'start' && lastCommand !== 'pause') {
-      this.log('startWebMonetization cancelled via', lastCommand)
-      this.activeTabLogger.log(
-        `startWebMonetization cancelled via ${lastCommand}`,
-        frame
-      )
-      return false
-    }
-
-    this.log('starting stream', requestId)
-    // We need to start this stream, even if we've already received a pause.
-    // That way we can "resume" it later.
-    this.assoc.setStreamId(frame, requestId)
-    this.assoc.setFrame(requestId, { tabId, frameId })
-    this.streams.beginStream(requestId, {
-      token: 'removeThis',
-      spspEndpoint,
-      ...request.data,
-      initiatingUrl: request.data.initiatingUrl
-    })
-
-    if (lastCommand === 'pause') {
-      // TODO: why do we need the timeout here ?
-      setTimeout(() => {
-        this.doPauseWebMonetization(frame, requestId)
-      }, 0)
-    }
-    return true
-  }
-
   /**
    * We can't allow these tip without an active stream because of the way
    * the tip events are using the destination currency.
    */
   async sendTip() {
     const tabId = this.activeTab
-    const streamId = this.assoc.getStreamId({ tabId, frameId: 0 })
+    const streamIds = this.assoc.getStreams({ tabId, frameId: 0 })
 
-    if (!streamId) return { success: false }
+    if (!streamIds.length) return { success: false }
+
+    const streamId = streamIds[0]
 
     try {
       const message = await this.tippingService.sendTip(
@@ -879,123 +555,8 @@ export class BackgroundScript {
     }
   }
 
-  private doPauseWebMonetization(frame: FrameSpec, requestId?: string) {
-    const id = requestId ?? this.assoc.getStreamId(frame)
-    if (id) {
-      this.tabStates.logLastMonetizationCommand(frame, 'pause', id)
-      this.log('pausing stream', id)
-      this.streams.pauseStream(id)
-      this.sendSetMonetizationStateMessage(frame, 'stopped')
-    }
-    return true
-  }
-
-  private doResumeWebMonetization(frame: FrameSpec, requestId?: string) {
-    const id = requestId ?? this.assoc.getStreamId(frame)
-    if (id) {
-      this.tabStates.logLastMonetizationCommand(frame, 'resume', id)
-      this.log('resuming stream', id)
-      this.sendSetMonetizationStateMessage(frame, 'pending')
-      this.streams.resumeStream(id)
-    }
-    return true
-  }
-
-  pauseWebMonetization(request: PauseWebMonetization, sender: MessageSender) {
-    if (this.tabStates.get(getTab(sender)).stickyState === 'sticky') {
-      return
-    }
-    return this.doPauseWebMonetization(
-      getFrameSpec(sender),
-      request.data.requestId
-    )
-  }
-
-  resumeWebMonetization(request: ResumeWebMonetization, sender: MessageSender) {
-    // Note that this gets sent regardless of whether actually monetized or not
-    // it's more like 'set tab interactive'
-    if (this.tabStates.get(getTab(sender)).playState === 'paused') {
-      return
-    }
-    return this.doResumeWebMonetization(
-      getFrameSpec(sender),
-      request.data.requestId
-    )
-  }
-
-  private handleStreamsAbortEvent() {
-    this.streams.on('abort', requestId => {
-      this.log('aborting monetization request', requestId)
-      const frame = this.assoc.getFrame(requestId)
-      if (frame) {
-        this.doStopWebMonetization(frame)
-      }
-    })
-  }
-
-  stopWebMonetization(sender: MessageSender) {
-    return this.doStopWebMonetization(getFrameSpec(sender))
-  }
-
-  private doStopWebMonetization(frame: FrameSpec) {
-    this.tabStates.setFrame(frame, { paymentPointer: undefined })
-
-    const requestId = this.assoc.getStreamId(frame)
-    if (requestId) {
-      this.tabStates.logLastMonetizationCommand(frame, 'stop', requestId)
-    }
-    const closed = this._closeStreams(frame.tabId, frame.frameId)
-    // May be noop other side if stop monetization was initiated from
-    // ContentScript
-    this.sendSetMonetizationStateMessage(frame, 'stopped', requestId)
-    if (closed) {
-      this.tabStates.clearFrame(frame)
-    }
-    this.reloadTabState({
-      from: 'stopWebMonetization'
-    })
-    return true
-  }
-
-  sendSetMonetizationStateMessage(
-    { tabId, frameId }: FrameSpec,
-    state: MonetizationState,
-    requestId?: string
-  ) {
-    const message: SetMonetizationState = {
-      command: 'setMonetizationState',
-      data: {
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        requestId: (this.assoc.getStreamId({ tabId, frameId }) ?? requestId)!,
-        state
-      }
-    }
-    this.api.tabs.sendMessage(tabId, message, { frameId })
-  }
-
   _closeStreams(tabId: number, frameId?: number) {
-    const streamIds = this.assoc.getTabStreams(tabId)
-    const haveFrameId = typeof frameId !== 'undefined'
-
-    let closed = 0
-    if (streamIds) {
-      Object.entries(streamIds).forEach(([innerFrameId, streamId]) => {
-        if (haveFrameId && Number(innerFrameId) !== frameId) {
-          return
-        }
-
-        this.log('closing stream with id', streamId)
-        this.streams.closeStream(streamId)
-        this.assoc.clearFrame(streamId)
-        closed++
-      })
-      if (haveFrameId) {
-        this.assoc.clearStream({ tabId, frameId: frameId as number })
-      } else {
-        this.assoc.clearTabStreams(tabId)
-      }
-    }
-    return !!closed
+    return this.monetization._closeStreams(tabId, frameId)
   }
 
   // This feature is no longer used
@@ -1007,18 +568,18 @@ export class BackgroundScript {
 
   private logout() {
     for (const tabId of this.tabStates.tabKeys()) {
-      // Make a copy as _closeStreams mutates and we want to actually close
+      // Make a copy as _closeStreams mutates, and we want to actually close
       // the streams before we set the state to stopped.
-      const requestIds = { ...this.assoc.getTabStreams(tabId) }
+      const requestIds = this.assoc.getTabStreams(tabId)
       this._closeStreams(tabId)
       this.tabStates.clear(tabId)
-      Object.entries(requestIds).forEach(([frameId, requestId]) => {
+      requestIds.forEach(([requestId, frame]) => {
         // TODO: this is hacky, but should do the trick for now
         const message: SetMonetizationState = {
           command: 'setMonetizationState',
           data: { state: 'stopped', requestId }
         }
-        this.api.tabs.sendMessage(tabId, message, { frameId: Number(frameId) })
+        this.api.tabs.sendMessage(tabId, message, { frameId: frame.frameId })
       })
     }
 
@@ -1034,46 +595,26 @@ export class BackgroundScript {
     })
     this.storage.clear()
     this.tabStates.setIcon(this.activeTab, 'unavailable')
-    this.reloadTabState()
+    this.tabStates.reloadTabState({ from: 'logout' })
     return true
   }
 
-  private setStreamControls(request: SetStreamControls, _: MessageSender) {
-    const tabId = this.activeTab
-    this.log('setStreamControls', request)
-
-    this.tabStates.set(tabId, {
-      stickyState: request.data.sticky,
-      playState: request.data.play
-    })
-    if (request.data.action === 'togglePlayOrPause') {
-      const tabState = this.tabStates.get(tabId)
-      const framesForTab = Object.keys(tabState.frameStates).map(Number)
-      this.log({ framesForTab })
-      if (request.data.play === 'paused') {
-        framesForTab.forEach(frameId => {
-          this.doPauseWebMonetization({ frameId, tabId })
-        })
-      } else if (request.data.play === 'playing') {
-        framesForTab.forEach(frameId => {
-          this.doResumeWebMonetization({ frameId, tabId })
-        })
-      }
-    }
-    this.reloadTabState({ from: request.command })
-    return true
-  }
-
-  private contentScriptInit(request: ContentScriptInit, sender: MessageSender) {
+  private async contentScriptInit(
+    request: ContentScriptInit,
+    sender: MessageSender
+  ): Promise<ContentScriptInitResponse> {
     const { tabId, frameId, spec } = getFrameSpec(sender)
     // Content script used to send a stopWebMonetization message every time
     // it loaded. Noop if no stream for tab.
     this._closeStreams(tabId, frameId)
     this.tabStates.clearFrame(spec)
-    this.reloadTabState({
+    this.tabStates.reloadTabState({
       from: 'onTabsUpdated status === contentScriptInit'
     })
-    return true
+    const wm2Allowed = await this.wm2OriginTrial.checkOrigin(
+      request.data.origin
+    )
+    return { wm2Allowed }
   }
 
   private async onFrameAllowedChanged(
