@@ -5,7 +5,7 @@ import * as tokens from '../../types/tokens'
 import { timeoutRejecting } from '../../util/timeout'
 
 /**
- * See {@link handleCoilTokenMessage}
+ * See {@link ContentAuthService#handleCoilTokenMessage}
  *
  * Refused to display 'https://coil.com/healthz' in a frame because an
  * ancestor violates the following Content Security Policy directive:
@@ -20,7 +20,7 @@ export class SiteToken {
     @inject(tokens.CoilDomain)
     private coilDomain: string,
     @inject(tokens.WextApi)
-    private api = chrome
+    private api: typeof chrome
   ) {}
 
   /**
