@@ -16,12 +16,11 @@ import {
   SPSPError,
   SPSPResponse
 } from '@webmonetization/polyfill-utils'
-import { Container, inject, injectable } from 'inversify'
+import { Container, inject, injectable, optional } from 'inversify'
 import { BandwidthTiers } from '@coil/polyfill-utils'
 
 import { notNullOrUndef } from '../../util/nullables'
 import * as tokens from '../../types/tokens'
-import { BTP_ENDPOINT } from '../../webpackDefines'
 import { BuildConfig } from '../../types/BuildConfig'
 
 import { AnonymousTokens } from './AnonymousTokens'
@@ -74,6 +73,12 @@ type OnMoneyEvent = {
   receipt?: string
 }
 
+interface CreateStreamDetails extends PaymentDetails {
+  token: string
+  spspEndpoint: string
+  initiatingUrl: string
+}
+
 @injectable()
 export class Stream extends EventEmitter {
   private readonly _requestId: string
@@ -104,20 +109,16 @@ export class Stream extends EventEmitter {
     @logger('Stream')
     private readonly _debug: Logger,
     private container: Container,
-    @inject(tokens.StreamDetails)
-    {
-      requestId,
-      spspEndpoint,
-      paymentPointer,
-      token,
-      initiatingUrl
-    }: PaymentDetails & {
-      token: string
-      spspEndpoint: string
-      initiatingUrl: string
-    }
+    @inject(tokens.CreateStreamDetails)
+    details: CreateStreamDetails,
+    @inject(tokens.BtpEndpoint)
+    @optional()
+    private btpEndpoint?: string
   ) {
     super()
+
+    const { requestId, spspEndpoint, paymentPointer, token, initiatingUrl } =
+      details
 
     this._paymentPointer = paymentPointer
     this._requestId = requestId
@@ -141,8 +142,8 @@ export class Stream extends EventEmitter {
     server.pathname = '/btp'
     this._server = server.href.replace(/^http/, 'btp+ws')
 
-    if (BTP_ENDPOINT) {
-      this._server = BTP_ENDPOINT
+    if (this.btpEndpoint) {
+      this._server = this.btpEndpoint
     }
   }
 
@@ -181,9 +182,10 @@ export class Stream extends EventEmitter {
       let btpToken: string | undefined
       let plugin, attempt
       try {
+        const spspDetailsPromise = await this._getSPSPDetails()
         btpToken = await this._anonTokens.getToken(this._authToken)
         plugin = await this._makePlugin(btpToken)
-        const spspDetails = await this._getSPSPDetails()
+        const spspDetails = await spspDetailsPromise
         this.container
           .rebind(tokens.NoContextLoggerName)
           .toConstantValue(`StreamAttempt:${this._requestId}:${++ATTEMPT}`)
@@ -248,12 +250,16 @@ export class Stream extends EventEmitter {
     this._debug('fetching spsp details. url=', spspUrl)
     let details: SPSPResponse
     try {
+      this.emit('spsp-event', 'loadstart', this._requestId)
       details = await getSPSPResponse(spspUrl, this._requestId)
+      this.emit('spsp-event', 'load', this._requestId)
     } catch (e) {
       if (e instanceof SPSPError) {
         const status = e.response?.status
+        this.emit('spsp-event', 'error', this._requestId, e.message)
         // Abort on Bad Request 4XX
         if (!status || (status >= 400 && status < 500)) {
+          this.emit('spsp-event', 'abort', this._requestId)
           this.abort()
         }
       }
