@@ -2,7 +2,7 @@ import { inject, interfaces, named } from 'inversify'
 
 import * as tokens from '../../types/tokens'
 
-import Context = interfaces.Context
+type Context = interfaces.Context
 
 export function logger(name?: string) {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
