@@ -1,27 +1,17 @@
-import React, { createContext, useContext, useState, useEffect } from 'react'
-import { StorageService } from '@webmonetization/wext/services'
+import React, { createContext, useContext, useEffect, useState } from 'react'
+import { StoreService } from '@webmonetization/wext/services'
 
 import { PopupHost } from '../types'
 import { getCreditCardFromPaymentMethods } from '../../util/getCreditCardFromPaymentMethods'
 import { calculateMaxAllowableTip } from '../../util/calculateMaxAllowableTip'
+import { StoreUpdate } from '../../types/commands'
 
-//
-// Models
-//
-export interface ITipContext {
-  currentTipAmountUsd: number
-  finalTipAmountUsd: number //* only used for the TipCompleteView so it renders the proper amount after local storage updates
-  maxAllowableTipAmountUsd: number //* the maxAllowableTip is primarily responsible for disabling tipping inputs
-  setCurrentTipAmountUsd: (amount: number) => void
-  setFinalTipAmountUsd: (amount: number) => void
-}
+import { ITipContext } from './iTipContext'
 
 interface ITipProvider {
-  storage: Pick<StorageService, 'get'>
+  storage: Pick<StoreService, 'get'>
   host: PopupHost
 }
-
-type StorageEventPartial = Pick<StorageEvent, 'key' | 'newValue'>
 
 //
 // Context
@@ -85,7 +75,7 @@ export const TipProvider: React.FC<ITipProvider> = props => {
 
     // need to update the values whenever the store user object is updated
     const events = host.events
-    events.on('storage', (evt: StorageEventPartial) => {
+    events.on('storeUpdate', (evt: StoreUpdate['data']) => {
       if (evt.key === 'user') {
         setValuesFromStorage()
       }
