@@ -1,8 +1,9 @@
 import ReactDOM from 'react-dom'
 import React from 'react'
+import { StorePersistence } from '@webmonetization/wext/services'
 
 import { API } from '../webpackDefines'
-import { StorageService } from '../services/storage'
+import { StoreService } from '../services/storage'
 import { ToPopupMessage } from '../types/commands'
 import { withSharedTheme } from '../shared-theme/withSharedTheme'
 import { openTab } from '../util/openTab'
@@ -10,20 +11,28 @@ import { openTab } from '../util/openTab'
 import { isExtension, mockPopupsPage } from './mocks/loadMockedStates'
 import { defaultPopupHost } from './context/popupHostContext'
 import { PopupHost } from './types'
-import { StorageEventPartial } from './context/storeContext'
 import { IndexWithRoot } from './Index'
-
-const isSafari = /^((?!chrome|android).)*safari/i.test(navigator.userAgent)
 
 export function run() {
   const rootEl = document.getElementById('root')
 
   if (isExtension) {
-    API.runtime.onMessage.addListener((message: ToPopupMessage) => {
-      if (message.command === 'closePopup') {
-        window.close()
+    const backgroundStore: StorePersistence = {
+      cache: new Map(),
+      clear(): void {
+        // noop, method never called popup side
+      },
+      removeItem(key: string): void {
+        API.runtime.sendMessage({ command: 'storeRemoveItem', data: { key } })
+      },
+      setItem(key: string, value): void {
+        API.runtime.sendMessage({
+          command: 'storeSetItem',
+          data: { key, value }
+        })
       }
-    })
+    }
+
     const host: PopupHost = {
       ...defaultPopupHost,
       runtime: {
@@ -31,40 +40,36 @@ export function run() {
         sendMessage: API.runtime.sendMessage.bind(API.runtime)
       }
     }
-    // TODO: this is fixed in Safari 15
-    window.addEventListener('storage', e => {
-      if (e.key === '$$popupCommand' && e.newValue) {
-        // Remove the timestamp which will cause a unique string used to trigger
-        // a `storage` event.
-        const command = e.newValue.substring(16)
-        const cmd: ToPopupMessage = JSON.parse(command)
-        if (cmd.command === 'closePopup') {
-          // window.close() itself actually causes a bad state on safari
-          if (isSafari) {
-            // eslint-disable-next-line no-console
-            console.warn(
-              'Should be running window.close(), ' +
-                'but it is buggy on safari, navigator.userAgent=' +
-                navigator.userAgent +
-                '\n' +
-                'see: ' +
-                'https://github.com/coilhq/web-monetization-projects/issues/1077'
-            )
-          } else {
-            window.close()
-          }
+
+    API.runtime.onMessage.addListener((message: ToPopupMessage) => {
+      if (message.command === 'closePopup') {
+        window.close()
+      } else if (message.command === 'storeUpdate') {
+        if (message.data.value) {
+          backgroundStore.cache.set(message.data.key, message.data.value)
+        } else {
+          backgroundStore.cache.delete(message.data.key)
         }
-      } else {
-        const event: StorageEventPartial = {
-          key: e.key,
-          newValue: e.newValue
-        }
-        host.events.emit('storage', event)
+        host.events.emit('storeUpdate', message.data)
       }
     })
-    ReactDOM.render(
-      <IndexWithRoot storage={new StorageService()} host={host} />,
-      rootEl
+
+    API.runtime.sendMessage(
+      {
+        command: 'storeGetItems'
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      },
+      (val: { items: Record<string, any> }) => {
+        backgroundStore.cache = new Map(Object.entries(val.items))
+
+        ReactDOM.render(
+          <IndexWithRoot
+            storage={new StoreService(backgroundStore)}
+            host={host}
+          />,
+          rootEl
+        )
+      }
     )
   } else {
     const MockPopupsPage = withSharedTheme(mockPopupsPage())
