@@ -1,8 +1,8 @@
-export interface BuildConfig {
+export interface BuildConfig extends Record<string, unknown> {
   logTabsApiEvents?: boolean
   extensionBuildString?: string
   extensionPopupFooterString?: string
   isCI?: boolean
-  isLoggingEnabled?: boolean
+  loggingEnabled?: boolean
   useLocalMockServer?: boolean
 }
