@@ -1,5 +1,4 @@
 import { User } from './user'
-import { PlayOrPauseState, StickyState } from './streamControls'
 
 export const STORAGE_KEY = {
   // BACKGROUND/POPUP
@@ -12,19 +11,14 @@ export const STORAGE_KEY = {
   monetizedTotal: 'monetizedTotal'
 }
 
-/**
- * All values are JSON encoded unless as stored
- * unless stated otherwise
- */
-export interface LocalStorageProxy {
+export interface StoreProxy {
+  // Auth State
   /**
    * The refresh token for authenticating to the coil site
-   * NOT JSON encoded
    */
   token?: string | null
 
   /**
-   * JSON serialized {@link User}
    */
   user?: User | null
   /**
@@ -32,18 +26,20 @@ export interface LocalStorageProxy {
    */
   validToken?: boolean | null
 
+  // Active Tab Monetization state
   adapted?: boolean | null
   monetized?: boolean | null
   coilSite?: string | null
-
-  stickyState?: StickyState | null
-  playState?: PlayOrPauseState | null
-
   monetizedTotal?: number | null
 
+  // State for popup builds
   extensionBuildString?: string | null
   extensionPopupFooterString?: string | null
 
+  // Popup state
   'popup-route:last'?: string | null
   'popup-route:tipping-shown'?: boolean | null
+
+  // Config
+  WM2_ALLOWED?: boolean | null
 }
