@@ -1,10 +1,14 @@
 export * from '@webmonetization/wext/tokens'
 export const CoilDomain = Symbol('CoilDomain')
-export const LocalStorageProxy = Symbol('LocalStorageProxy')
+export const BtpEndpoint = Symbol('BtpEndpoint')
+export const StoreProxy = Symbol('StoreProxy')
 export const ContentRuntime = Symbol('ContentRuntime')
 export const Logger = Symbol('Logger')
 export const NoContextLoggerName = Symbol('Logger')
-export const StreamDetails = Symbol('StreamDetails')
+export const CreateStreamDetails = Symbol('CreateStreamDetails')
 export const BuildConfig = Symbol('BuildConfig')
+export const ActiveTab = Symbol('ActiveTab')
 export const LoggingEnabled = Symbol('LoggingEnabled')
 export const UserAgent = Symbol('UserAgent')
+export const TokenStore = Symbol('TokenStore')
+export const StorePersistence = Symbol('StorePersistence')
