@@ -1,8 +1,8 @@
 import * as asyncUtils from './lib/asyncUtils'
 
-export { MonetizationTagObserver } from './lib/MonetizationTagObserver'
+export { MonetizationTagManager } from './lib/MonetizationTagManager'
 
-export type { PaymentDetails } from './lib/MonetizationTagObserver'
+export type { PaymentDetails } from './lib/MonetizationTagManager'
 
 export { watchMouseMovement, watchVisibility } from './lib/idle'
 export { whenDocumentReady } from './lib/whenDocumentReady'
@@ -15,7 +15,11 @@ export { AdaptiveBandwidth } from './lib/AdaptiveBandwidth'
 export type { StreamControl } from './lib/idle'
 export { watchPageEvents } from './lib/idle'
 export { BackoffWaiter } from './lib/BackoffWaiter'
-export { resolvePaymentEndpoint } from './lib/resolvePaymentEndpoint'
+export {
+  resolvePaymentEndpoint,
+  PaymentEndpointError
+} from './lib/resolvePaymentEndpoint'
 export { asyncUtils }
 export { getFarFutureExpiry } from './lib/getFarFutureExpiry'
 export { parsePolicyDirectives } from './lib/parsePolicyDirectives'
+export { mozClone } from './lib/mozClone'
