@@ -1,11 +1,14 @@
+import { wm2Polyfill } from './wm2Polyfill'
+
 const createBindingCode = (...events: string[]) => {
+  const colorCode = `color: aqua; background-color: black`
   return events
     .map(
       e =>
         `document.monetization.addEventListener('${e}', ` +
         `(e) => console.log(
-           '%c Web-Monetization %s event:  %s',
-           'color: aqua; background-color: black',
+           '%c WM1 %s event:  %s',
+           '${colorCode}',
            e.type,
            JSON.stringify(e.detail)) )
           `
@@ -24,11 +27,20 @@ const basicEventsLoggingCode = createBindingCode(
 const progressLoggingCode = createBindingCode('monetizationprogress')
 
 // language=JavaScript
-export const wmPolyFillMinimal = `
-  document.monetization = document.createElement('div')
-  document.monetization.state = 'stopped'
+export const wm1Polyfill = `
+  class Monetization extends EventTarget {
+    state = 'stopped'
+
+    get [Symbol.toStringTag]() {
+      return 'Monetization'
+    }
+  }
+
+  document.monetization = new Monetization()
+  // TODO: use coil-monetization-v1 and update content script that sends
+  // the events
   document.addEventListener('monetization-v1', function(event) {
-    const {type, detail} = event.detail
+    const { type, detail } = event.detail
     if (type === 'monetizationstatechange') {
       document.monetization.state = detail.state
     } else {
@@ -42,7 +54,7 @@ export const wmPolyFillMinimal = `
 
 // language=JavaScript
 export const wmPolyfill = `
-  ${wmPolyFillMinimal}
+  ${wm1Polyfill}
 
   if (localStorage.WM_DEBUG) {
     ${basicEventsLoggingCode}
@@ -51,10 +63,20 @@ export const wmPolyfill = `
   if (localStorage.WM_DEBUG_PROGRESS) {
     ${progressLoggingCode}
   }
+
+  // The extension tests origin of requests from content scripts, this is just
+  // so that people do not get confused by the presence of WM2 polyfill objects
+  // when the requests will be blocked due to not being part of the origin
+  // trial.
+  // Block scoped so vars will not be set on window unless explicitly done
+  if (document.head.dataset['wm2Allowed'] === 'true') {
+    delete document.head.dataset['wm2Allowed']
+    ${wm2Polyfill}
+  }
 `
 
 export const includePolyFillMessage = `
 Unable to inject \`document.monetization\` polyfill!
 Include the polyfill in your page:
-<script type="application/javascript">${wmPolyFillMinimal}</script>
+<script type="application/javascript">${wm1Polyfill}</script>
 `
